+++ src/DeprecatedUuidMethodsTrait.php
@@ -66,16 +66,6 @@ trait DeprecatedUuidMethodsTrait
     /**
      * @deprecated Use {@see UuidInterface::getFields()} to get a
      *     {@see FieldsInterface} instance. If it is a {@see Rfc4122FieldsInterface}
-     *     instance, you may call {@see Rfc4122FieldsInterface::getClockSeqHiAndReserved()}.
-     */
-    public function getClockSeqHiAndReservedHex(): string
-    {
-        return $this->fields->getClockSeqHiAndReserved()->toString();
-    }
-
-    /**
-     * @deprecated Use {@see UuidInterface::getFields()} to get a
-     *     {@see FieldsInterface} instance. If it is a {@see Rfc4122FieldsInterface}
      *     instance, you may call {@see Rfc4122FieldsInterface::getClockSeqLow()}
      *     and use the arbitrary-precision math library of your choice to
      *     convert it to a string integer.